# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ITU API",
    "description": "This API exposes data about: ITU students and enrollments;  course workloads for each student; course teachers, and their student popularity scores and student satisfaction rating for courses",
    "version": "1.0.0"
  },
  "host": "api.example.com",
  "basePath": "/v1",
  "schemes": [
    "https"
  ],
  "paths": {
    "/Student": {
      "post": {
        "summary": "Creates a student",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Student",
            "description": "Student object that needs to be added",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Student"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created student"
          }
        }
      }
    },
    "/Student/{StudentID}": {
      "put": {
        "summary": "Update existing student",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "StudentID",
            "in": "path",
            "description": "Student object that needs to be updated given by id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Student updated",
            "schema": {
              "$ref": "#/definitions/Student"
            }
          },
          "404": {
            "description": "Student ID not found"
          }
        }
      },
      "get": {
        "parameters": [
          {
            "name": "StudentID",
            "in": "path",
            "required": true,
            "description": "Returns a student with given id",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "A single student",
            "schema": {
              "$ref": "#/definitions/Course"
            }
          },
          "400": {
            "description": "Invalid input type"
          },
          "404": {
            "description": "Student ID not found"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "name": "StudentID",
            "in": "path",
            "required": true,
            "description": "Deletes the student with given id",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Student deleted"
          },
          "400": {
            "description": "Invalid input type"
          },
          "404": {
            "description": "Student ID not found"
          }
        }
      }
    },
    "/Course": {
      "post": {
        "summary": "Creates a course",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Course",
            "description": "A Course is something offered by the university. A course has a teacher, students taking the course and satisfaction rating ",
            "schema": {
              "$ref": "#/definitions/Course"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Course/{CourseID}": {
      "put": {
        "summary": "Update existing course",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "CourseID",
            "in": "path",
            "description": "Course object that needs to be added",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Course updated",
            "schema": {
              "$ref": "#/definitions/Course"
            }
          },
          "404": {
            "description": "Course ID not found"
          }
        }
      },
      "get": {
        "parameters": [
          {
            "name": "CourseID",
            "in": "path",
            "required": true,
            "description": "Returns a course with given id",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "A single course",
            "schema": {
              "$ref": "#/definitions/Course"
            }
          },
          "400": {
            "description": "Invalid input type"
          },
          "404": {
            "description": "Course ID not found"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "name": "CourseID",
            "in": "path",
            "required": true,
            "description": "Deletes the course with given id",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Course deleted"
          },
          "400": {
            "description": "Invalid input type"
          },
          "404": {
            "description": "Course ID not found"
          }
        }
      }
    },
    "/Teacher": {
      "post": {
        "summary": "Creates a Teacher",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Course",
            "description": "A teacher is someone who is teaching at the university. A Teacher has a name, an id and an overall student popularity score.",
            "schema": {
              "$ref": "#/definitions/Teacher"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Teacher/{TeacherID}": {
      "put": {
        "summary": "Update existing teacher",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "TeacherID",
            "in": "path",
            "description": "Teacher object that needs to be added",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Teacher updated",
            "schema": {
              "$ref": "#/definitions/Teacher"
            }
          },
          "400": {
            "description": "Invalid input type"
          },
          "404": {
            "description": "Teacher ID not found"
          }
        }
      },
      "get": {
        "parameters": [
          {
            "name": "TeacherID",
            "in": "path",
            "required": true,
            "description": "Returns a teacher with given id",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "A single teacher",
            "schema": {
              "$ref": "#/definitions/Teacher"
            }
          },
          "400": {
            "description": "Invalid input type"
          },
          "404": {
            "description": "Teacher ID not found"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "name": "TeacherID",
            "in": "path",
            "required": true,
            "description": "Deletes the teacher with given id",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Teacher deleted"
          },
          "400": {
            "description": "Invalid input type"
          },
          "404": {
            "description": "Teacher ID not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Student": {
      "type": "object",
      "required": [
        "name",
        "age",
        "id",
        "enrollment"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "age": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "enrollment": {
          "type": "boolean"
        }
      },
      "example": {
        "name": "Bob",
        "age": 42,
        "id": 1,
        "enrollment": true
      }
    },
    "Course": {
      "type": "object",
      "required": [
        "teacher",
        "id",
        "name",
        "etcs",
        "rating"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "teacher": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "rating": {
          "type": "number"
        },
        "etcs": {
          "type": "number"
        }
      }
    },
    "Teacher": {
      "type": "object",
      "required": [
        "name",
        "id",
        "score"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "score": {
          "type": "number"
        }
      },
      "example": {
        "name": "Alice",
        "id": "25",
        "score": 5.4
      }
    }
  }
}